+++ enterprise/frontend/src/metabase-enterprise/whitelabel/components/FontFileSettings/FontFileSettings.tsx
@@ -1,7 +1,8 @@
 import React, { FocusEvent, useCallback, useMemo } from "react";
 import { t } from "ttag";
 import Input from "metabase/core/components/Input";
-import { FontFile, FontFileOption } from "./types";
+import { FontFile } from "metabase-types/api";
+import { FontFileOption } from "./types";
 import { FONT_OPTIONS, getFontFiles, getFontUrls } from "./utils";
 import {
   TableBody,
