+++ enterprise/frontend/src/metabase-enterprise/whitelabel/components/FontFileSettings/types.ts
@@ -1,12 +1,4 @@
-export interface FontFile {
-  src: string;
-  fontWeight: number;
-  fontFormat: FontFormat;
-}
-
 export interface FontFileOption {
   name: string;
   fontWeight: number;
 }
-
-export type FontFormat = "woff" | "woff2" | "truetype";
