+++ enterprise/frontend/src/metabase-enterprise/whitelabel/components/FontFileSettings/utils.ts
@@ -1,6 +1,7 @@
 import { t } from "ttag";
 import _ from "underscore";
-import { FontFile, FontFileOption, FontFormat } from "./types";
+import { FontFile, FontFormat } from "metabase-types/api";
+import { FontFileOption } from "./types";
 
 export const FONT_OPTIONS: FontFileOption[] = [
   {
