+++ frontend/src/metabase-types/api/settings.ts
@@ -1,17 +1,25 @@
-export interface EngineSourceContact {
-  name?: string;
-  address?: string;
+export interface Engine {
+  "driver-name": string;
+  "superseded-by": string | undefined;
+  source: EngineSource;
 }
 export interface EngineSource {
   type?: "official" | "community" | "partner";
   contact?: EngineSourceContact;
 }
 
-export interface Engine {
-  "driver-name": string;
-  "superseded-by": string | undefined;
-  source: EngineSource;
+export interface EngineSourceContact {
+  name?: string;
+  address?: string;
 }
+
+export interface FontFile {
+  src: string;
+  fontWeight: number;
+  fontFormat: FontFormat;
+}
+
+export type FontFormat = "woff" | "woff2" | "truetype";
 
 export interface Version {
   tag: string;
@@ -20,6 +28,7 @@ export interface Version {
 export type LocaleData = [string, string];
 
 export interface Settings {
+  "application-font-files": FontFile[] | undefined;
   "available-locales": LocaleData[] | undefined;
   "enable-public-sharing": boolean;
   "enable-xrays": boolean;
